+++ src/components/DefaultInput.tsx
@@ -1,19 +1,22 @@
 import clsx from 'clsx';
-import React, { InputHTMLAttributes } from 'react';
+import React, { InputHTMLAttributes, Ref, forwardRef } from 'react';
 import './DefaultInput.css';
 
-const DefaultInput = (
-  props: InputHTMLAttributes<HTMLInputElement>
-): JSX.Element => {
-  const { className, ...rest } = props;
+const DefaultInput = forwardRef(
+  (
+    props: InputHTMLAttributes<HTMLInputElement>,
+    ref: Ref<HTMLInputElement>
+  ): JSX.Element => {
+    const { className, ...rest } = props;
 
-  return (
-    <input
-      className={clsx('react-loqate-input', className)}
-      data-testid="default-input"
-      {...rest}
-    />
-  );
-};
+    return (
+      <input
+        className={clsx('react-loqate-input', className)}
+        ref={ref}
+        {...rest}
+      />
+    );
+  }
+);
 
 export default DefaultInput;