+++ src/components/DefaultListItem.tsx
@@ -6,7 +6,7 @@ import './DefaultListItem.css';
 const DefaultListItem = (
   props: LiHTMLAttributes<HTMLLIElement> & { suggestion: Item }
 ): JSX.Element => {
-  const { className, value, children, suggestion, ...rest } = props;
+  const { className, value, suggestion, onKeyDown, ...rest } = props;
 
   const regex = new RegExp(
     `(.*)(${value?.toString().replace(/\W/g, '')})(.*)`,
@@ -21,8 +21,29 @@ const DefaultListItem = (
   return (
     <li
       className={clsx('react-loqate-list-item', className)}
-      aria-label={match}
-      tabIndex={0}
+      aria-label={before + match + after}
+      onKeyDown={(e) => {
+        onKeyDown?.(e);
+
+        if (e.key === 'ArrowDown') {
+          // don't scroll the page
+          e.preventDefault();
+          const listItem = e.target as HTMLLIElement;
+          const next = listItem.nextSibling as HTMLLIElement;
+          if (next) {
+            next.focus();
+          }
+        }
+        if (e.key === 'ArrowUp') {
+          // don't scroll the page
+          e.preventDefault();
+          const listItem = e.target as HTMLLIElement;
+          const previous = listItem.previousSibling as HTMLLIElement;
+          if (previous) {
+            previous.focus();
+          }
+        }
+      }}
       {...rest}
     >
       {result?.length === 4 ? (