+++ src/utils/__tests__/Loqate.test.ts
@@ -1,15 +1,12 @@
-import Loqate from '../Loqate';
-import { server } from '../../__tests__/server';
+import { fetch } from 'cross-fetch';
+import { describe, expect, it } from 'vitest';
 import { selection } from '../../__tests__/__fixtures__/selection';
 import { suggestions } from '../../__tests__/__fixtures__/suggestions';
+import { server } from '../../__tests__/server';
 import { errorHandler } from '../../__tests__/serverHandlers';
-import { describe, beforeAll, afterEach, afterAll, it, expect } from 'vitest';
-import { fetch } from 'cross-fetch';
+import Loqate from '../Loqate';
 
 global.fetch = fetch;
-beforeAll(() => server.listen({ onUnhandledRequest: 'error' }));
-afterAll(() => server.close());
-afterEach(() => server.resetHandlers());
 
 describe('Loqate', () => {
   it('should initialize', () => {
@@ -44,7 +41,7 @@ describe('Loqate', () => {
       expect({ Items }).toEqual(suggestions);
     });
 
-    it('should throw loqate errors', async () => {
+    it('should throw errors', async () => {
       server.use(errorHandler);
 
       const loqate = Loqate.create('some-key');
@@ -57,10 +54,36 @@ describe('Loqate', () => {
           limit: 10,
           containerId: 'some-container-id',
         });
-      }).rejects.toThrowError(
-        new Error(
-          'Loqate error: {"Error":"2","Description":"Unknown key","Cause":"The key you are using to access the service was not found.","Resolution":"Please check that the key is correct. It should be in the form AA11-AA11-AA11-AA11."}'
-        )
+      }).rejects.toThrowError(new Error('Unknown key'));
+    });
+
+    it('should throw loqate errors', async () => {
+      server.use(errorHandler);
+
+      const loqate = Loqate.create('some-key');
+
+      let error;
+      try {
+        await loqate.find({
+          text: 'some-text',
+          language: 'some-language',
+          countries: ['GB', 'US'],
+          limit: 10,
+          containerId: 'some-container-id',
+        });
+      } catch (e) {
+        error = e;
+      }
+
+      expect(error).toEqual(new Error('Unknown key'));
+      expect(JSON.stringify(error)).toEqual(
+        JSON.stringify({
+          Cause: 'The key you are using to access the service was not found.',
+          Description: 'Unknown key',
+          Error: '2',
+          Resolution:
+            'Please check that the key is correct. It should be in the form AA11-AA11-AA11-AA11.',
+        })
       );
     });
   });